+++ src/main/java/com/example/model/inter/InterZobowiazania.java
@@ -14,7 +14,7 @@ import java.util.List;
 public interface InterZobowiazania {
 
  List<Zobowiazania> getAll(); 
-  Zobowiazania getById(int id);
+  Zobowiazania getById(Long id);
  void add(Zobowiazania zobowiazania); 
  void delete(Long id); 
  void update(Zobowiazania zobowiazania);
